Move server-related code Action Cable from connection/ to server/
