+++ actioncable/lib/action_cable/server/base.rb
@@ -69,7 +69,7 @@ module ActionCable
       end
 
       def event_loop
-        @event_loop || @mutex.synchronize { @event_loop ||= ActionCable::Connection::StreamEventLoop.new }
+        @event_loop || @mutex.synchronize { @event_loop ||= StreamEventLoop.new }
       end
 
       # The worker pool is where we run connection callbacks and channel actions. We
