+++ actioncable/lib/action_cable/server/connection.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module ActionCable
+  module Server
+    class Connection
+    end
+  end
+end
