+++ actioncable/lib/action_cable/server/connection/web_socket.rb
@@ -1,0 +1,45 @@
+# frozen_string_literal: true
+
+require "websocket/driver"
+
+module ActionCable
+  module Server
+    class Connection
+      # = Action Cable \Connection \WebSocket
+      #
+      # Wrap the real socket to minimize the externally-presented API
+      class WebSocket # :nodoc:
+        def initialize(env, event_target, event_loop, protocols: ActionCable::INTERNAL[:protocols])
+          @websocket = ::WebSocket::Driver.websocket?(env) ? ClientSocket.new(env, event_target, event_loop, protocols) : nil
+        end
+
+        def possible?
+          websocket
+        end
+
+        def alive?
+          websocket&.alive?
+        end
+
+        def transmit(...)
+          websocket&.transmit(...)
+        end
+
+        def close(...)
+          websocket&.close(...)
+        end
+
+        def protocol
+          websocket&.protocol
+        end
+
+        def rack_response
+          websocket&.rack_response
+        end
+
+        private
+          attr_reader :websocket
+      end
+    end
+  end
+end
