+++ actioncable/test/stubs/test_server.rb
@@ -31,7 +31,7 @@ class TestServer
   end
 
   def event_loop
-    @event_loop ||= ActionCable::Connection::StreamEventLoop.new.tap do |loop|
+    @event_loop ||= ActionCable::Server::StreamEventLoop.new.tap do |loop|
       loop.instance_variable_set(:@executor, Concurrent.global_io_executor)
     end
   end
